pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "droasms-bot"
version = "0.4.0"
description = "Discord bot economy foundations."
readme = "README.md"
requires-python = ">=3.13"
license = { file = "LICENSE" }
authors = [
    { name = "DRoASMS Team" },
]
dependencies = [
    "discord.py>=2.4.0,<3.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "alembic>=1.13.0,<2.0.0",
    "SQLAlchemy>=2.0.30,<3.0.0",
    "structlog>=24.1.0,<25.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<0.24.0",
    "mypy>=1.11.0,<2.0.0",
    "ruff>=0.6.0,<0.7.0",
    "black>=24.8.0,<25.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.build.targets.sdist]
include = ["src", "tests", "README.md", "LICENSE"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "ASYNC",
    "B",
    "C4",
]

[tool.ruff.lint.per-file-ignores]
"src/db/migrations/versions/*.py" = ["E501", "I001"]
"src/db/seeds/*.py" = ["E501"]
"tests/**/*.py" = ["E501"]
"src/bot/commands/state_council.py" = ["E501"]

[tool.black]
line-length = 100
target-version = ["py313"]
include = "\\.pyi?$"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]
markers = [
    "performance: Performance and load tests for NFR validation",
]

[tool.mypy]
python_version = "3.13"
namespace_packages = true
files = ["src"]
explicit_package_bases = true
enable_error_code = "ignore-without-code"
strict = true

[[tool.mypy.overrides]]
module = [
    "asyncpg",
    "asyncpg.*",
]
ignore_missing_imports = true
